Horde: Move JobStepState, JobStepOutcome and Priority enums into native C# code (rather than protobuf enums)
#rnx

[CL 32780416 by ben marsh in ue5-main branch]

+++ Engine/Source/Programs/Horde/Horde.Agent/Execution/JobExecutor.cs
@@ -669,7 +669,7 @@ namespace Horde.Agent.Execution
 					}
 					createNode.RunEarly = exportedNode.RunEarly;
 					createNode.Warnings = exportedNode.Warnings;
-					createNode.Priority = Priority.Normal;
+					createNode.Priority = (int)Priority.Normal;
 					createNode.Annotations.Add(exportedNode.Annotations);
 					createGroup.Nodes.Add(createNode);
 				}
